Update box.rb
this @rye_stty_save variable doesn't appear to be used anywhere else but actually causes a problem when using rye as part of a non interactive terminal session (like build tools/deployer tools) - using the stty command on these sessions will result in an error (as no terminal is present) and a lot of noise in the output (it will generate an error on every connection). 

If it's not actively being used for anything I'd rather remove it.